requirements.txt
fastapi
uvicorn
openai
pydantic
textblob
python-dotenv  
